feat: resolved comments
